requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# python=3.6.10
# numpy=1.15.4
# pillow=5.4.1
# pip install pydensecrf==1.0rc3
# pytorch=1.3.1
# pyyaml=3.13
# torchvision=0.4.2
# tqdm==4.31.1
# pip install tensorboardX==1.6
# packaging=19.0
# scikit-learn=0.20.2
# pip install scikit-image==0.16.1
# six=1.12.0
# scipy=1.2.1

# python=3.6.3
numpy=1.15.4
pillow=5.4.1
pydensecrf=1.0rc3
pytorch=1.0.1
pyyaml=3.13
torchvision=0.2.1
tqdm==4.31.1
tensorboardX=1.6
packaging=19.0
scikit-learn=0.20.2
scikit-image=0.16.1
six=1.12.0
scipy=1.2.1
